refactor(#360): lift the YAML serializer out of editStore into lib/serializePipeline
Tranche 1 of #360. Verbatim move, zero logic change, no byte of emitted YAML
differs. `editStore.ts:255-537` (283 lines: the canonical pipeline YAML emitter
plus the node-library emitter from #345) becomes
`frontend/src/lib/serializePipeline.ts`, a pure leaf whose only dependency is a
fully-erased `import type` from ../types.

Why: those 283 lines were never a store. They are the single code path producing
the bytes written to `.pdo/pipelines/*.yaml`, the node-library export, and both
sides of the library synced/diverged star comparison. Keeping them in the store
layer forced `lib/layoutFields.test.ts` to reach up into `stores/` for the
serializer it exists to police -- the only lib->stores back-edge in src/lib (22
modules). That edge is now gone; the emit/strip couple (`pipelineToYamlObject` +
`stripLayout`) sits as two siblings instead of straddling three directories.

editStore.ts 1409 -> 1126 lines. `PortDef`/`PortSide` dropped from its type
imports (their only use left with `portToYamlObject`; `noUnusedLocals` would
have failed `tsc`). One remaining call site: `editStore.save`.

Hard cut, no re-export shim -- nothing flags an exported-but-unused symbol, so a
shim would rot invisibly. All 8 importers re-pointed (4 were split imports).

Tests: the 33 pure-serializer tests move to `lib/serializePipeline.test.ts` with
their local fixture helpers. The 4 store-coupled cases that merely assert on YAML
(notes round-trip, the `addEdge`-minted region id, `updateRegion`, and the
"via save" trio reading the api mock) stay in `editStore.test.ts` -- they are
store tests, not serializer tests.

Test-count identity, the load-bearing gate here because vitest is NOT run in CI
(ci.yml stops at typecheck + lint + build, so all 1324 unit tests are non-gating
and a dropped describe block would go unnoticed):
  before: 96 files / 1324 tests    after: 97 files / 1324 tests
  editStore 125 + serializePipeline 33 + layoutFields 11 = 169
Same total, one more file => 33 tests relocated, none lost.

The two per-port emitters were left DELIBERATELY divergent: the pipeline emit
writes `side` whenever set, the node-library emit omits it when it equals the
direction default (round-trips by absence). They now sit ~100 lines apart in one
module and read like duplication. Unifying them would strip `side:` from every
pipeline file and flip every library star to `diverged`. A comment on
`portToYamlObject` says so.

Also recorded as a module docstring: everything this file emits is parsed by the
daemon's serde-derived reader, and `exportNodeAsYaml` must match
`library_store::LibraryEntry` field-for-field -- a cross-language contract no
test crosses (#352 is what that gap cost). Not fixed here; worth its own issue.

Docs: 2 stale prose pointers re-pointed (lib/layoutFields.ts, lib/edgePath.ts)
and CONTEXT.md:170 given the serializer's new path. Both prose sites also
corrected a false claim -- they said an unclassified new field "fails the build";
it fails `layoutFields.test.ts` under vitest, which CI never runs.

Gate, all green after every edit: tsc -b 0, eslint 0, vitest 97/1324,
npm run build 0, net line delta +41 (= 26-line docstring + 6-line port note +
4-line test header + 4 split-import lines + 1 blank).

Tranche 2 (the tabsStore split) is NOT in this commit and stays ready-for-human:
22 non-test files bind `useEditStore` and ADR-0014's undo-COW invariant is
load-bearing there. #360 stays open.
