+++ frontend/src/stores/editStore.test.ts
@@ -1,5 +1,6 @@
 import { describe, it, expect, beforeEach, afterEach, vi } from "vitest";
-import { useEditStore, serializePipeline, exportNodeAsYaml } from "./editStore";
+import { useEditStore } from "./editStore";
+import { serializePipeline } from "../lib/serializePipeline";
 import type { OpenPipeline, Selection } from "./editStore";
 import { generatedRegionId } from "../lib/loopRegions";
 import { pipelinesEquivalent } from "../hooks/useLibraryPipelines";
@@ -1295,487 +1296,6 @@ describe("createCollectionRegion (#151 / #269)", () => {
   });
 });
 
-describe("serializePipeline round-trip: YAML structural correctness", () => {
-  function makeFullPipeline(extraNodes: NodeDef[], edges: EdgeDef[] = []): PipelineDef {
-    const start: NodeDef = {
-      id: "start", name: "Start", type: "start",
-      inputs: [], outputs: [{ name: "user_prompt", repeated: false, side: "right" }],
-      interactive: false, view: { x: 0, y: 0 },
-    };
-    const end: NodeDef = {
-      id: "end", name: "End", type: "end",
-      inputs: [{ name: "result", repeated: false, side: "left" }], outputs: [],
-      interactive: false, view: { x: 400, y: 0 },
-    };
-    return {
-      name: "round-trip-test", version: "1.0", variables: {},
-      nodes: [start, ...extraNodes, end], edges,
-    };
-  }
-
-  it("serializes a minimal start+end pipeline to parseable YAML", () => {
-    const pipeline = makeFullPipeline([]);
-    const yaml = serializePipeline(pipeline);
-    expect(yaml).toContain("name: round-trip-test");
-    expect(yaml).toContain("type: start");
-    expect(yaml).toContain("type: end");
-  });
-
-  it("serializes a bounded loops: region block (ADR-0011 / #148)", () => {
-    const impl: NodeDef = {
-      id: "impl", name: "implementer", type: "code-mutating",
-      inputs: [], outputs: [{ name: "code", repeated: false, side: "right" }],
-      interactive: false, view: { x: 200, y: 0 },
-    };
-    const rev: NodeDef = {
-      id: "rev", name: "reviewer", type: "doc-only",
-      inputs: [], outputs: [{ name: "review", repeated: false, side: "right" }],
-      interactive: false, view: { x: 300, y: 0 },
-    };
-    const pipeline = makeFullPipeline([impl, rev]);
-    pipeline.loops = [
-      { id: "review_loop", kind: "bounded", members: ["impl", "rev"], max_iter: 3 },
-    ];
-    const yaml = serializePipeline(pipeline);
-    expect(yaml).toContain("loops:");
-    expect(yaml).toContain("id: review_loop");
-    expect(yaml).toContain("kind: bounded");
-    expect(yaml).toContain("max_iter: 3");
-    // members listed
-    expect(yaml).toMatch(/members:/);
-  });
-
-  it("serializes a collection loops: region with its over: field (#269)", () => {
-    const worker: NodeDef = {
-      id: "worker", name: "worker", type: "code-mutating",
-      inputs: [], outputs: [{ name: "out", repeated: false, side: "right" }],
-      interactive: false, view: { x: 200, y: 0 },
-    };
-    const pipeline = makeFullPipeline([worker]);
-    pipeline.loops = [
-      { id: "fan_out", kind: "collection", members: ["worker"], over: "items" },
-    ];
-    const yaml = serializePipeline(pipeline);
-    expect(yaml).toContain("loops:");
-    expect(yaml).toContain("id: fan_out");
-    expect(yaml).toContain("kind: collection");
-    expect(yaml).toContain("over: items");
-  });
-
-  it("omits the loops: block when there are no regions", () => {
-    const yaml = serializePipeline(makeFullPipeline([]));
-    expect(yaml).not.toContain("loops:");
-  });
-
-  it("emits prompt_required: false for prompt-optional pipelines (#158)", () => {
-    const pipeline = makeFullPipeline([]);
-    pipeline.prompt_required = false;
-    const yaml = serializePipeline(pipeline);
-    expect(yaml).toContain("prompt_required: false");
-  });
-
-  it("omits prompt_required when prompt-required (the default, #158)", () => {
-    const requiredExplicit = makeFullPipeline([]);
-    requiredExplicit.prompt_required = true;
-    expect(serializePipeline(requiredExplicit)).not.toContain("prompt_required");
-
-    // Absent flag is the prompt-required default → still omitted.
-    const absent = makeFullPipeline([]);
-    expect(serializePipeline(absent)).not.toContain("prompt_required");
-  });
-
-  it("emits a per-node model override when set (#296)", () => {
-    const impl: NodeDef = {
-      id: "impl", name: "implementer", type: "code-mutating",
-      inputs: [], outputs: [{ name: "code", repeated: false, side: "right" }],
-      interactive: false, view: { x: 200, y: 0 }, model: "opus",
-    };
-    const yaml = serializePipeline(makeFullPipeline([impl]));
-    expect(yaml).toContain("model: opus");
-  });
-
-  it("omits model when unset — the byte-identical / no-diverge default (#296)", () => {
-    const impl: NodeDef = {
-      id: "impl", name: "implementer", type: "code-mutating",
-      inputs: [], outputs: [{ name: "code", repeated: false, side: "right" }],
-      interactive: false, view: { x: 200, y: 0 },
-    };
-    expect(serializePipeline(makeFullPipeline([impl]))).not.toContain("model:");
-  });
-
-  it("serializes output port with frontmatter at correct indentation", () => {
-    const reviewer: NodeDef = {
-      id: "reviewer", name: "reviewer", type: "doc-only",
-      inputs: [{ name: "code", repeated: false, side: "left" }],
-      outputs: [{
-        name: "review", repeated: false, side: "right",
-        frontmatter: {
-          verdict: { type: "enum", allowed: ["PASS", "FAIL"] },
-        },
-      }],
-      interactive: false, view: { x: 200, y: 0 },
-    };
-    const yaml = serializePipeline(makeFullPipeline([reviewer]));
-
-    // The frontmatter fields (type/allowed) must be siblings, not parent-child
-    const lines = yaml.split("\n");
-    const typeIdx = lines.findIndex((l) => l.includes("type: enum"));
-    const allowedIdx = lines.findIndex((l) => l.includes("allowed:"));
-    expect(typeIdx).toBeGreaterThan(-1);
-    expect(allowedIdx).toBeGreaterThan(-1);
-
-    // Both should have the same leading whitespace (they're siblings under verdict:)
-    const typeIndent = lines[typeIdx].match(/^(\s*)/)?.[1].length ?? -1;
-    const allowedIndent = lines[allowedIdx].match(/^(\s*)/)?.[1].length ?? -1;
-    expect(typeIndent).toBe(allowedIndent);
-  });
-
-  it("serializes an edge when clause at correct indentation", () => {
-    const gate: NodeDef = {
-      id: "gate", name: "gate", type: "doc-only",
-      inputs: [{ name: "in", repeated: false, side: "left" }],
-      outputs: [{ name: "out", repeated: false, side: "right" }],
-      interactive: false, view: { x: 200, y: 0 },
-    };
-    const yaml = serializePipeline(
-      makeFullPipeline([gate], [
-        {
-          source: { node: "gate", port: "out" },
-          target: { node: "end", port: "result" },
-          when: { verdict: { eq: "PASS" }, score: { gte: 7 } },
-        },
-      ]),
-    );
-
-    const lines = yaml.split("\n");
-    // Find verdict and score lines under when: — they must be at same indent
-    const verdictIdx = lines.findIndex((l) => l.includes("verdict:"));
-    const scoreIdx = lines.findIndex((l) => l.includes("score:"));
-    expect(verdictIdx).toBeGreaterThan(-1);
-    expect(scoreIdx).toBeGreaterThan(-1);
-
-    const verdictIndent = lines[verdictIdx].match(/^(\s*)/)?.[1].length ?? -1;
-    const scoreIndent = lines[scoreIdx].match(/^(\s*)/)?.[1].length ?? -1;
-    expect(verdictIndent).toBe(scoreIndent);
-  });
-
-  it("serializes a manual edge's mode and waypoints (shareable routing, #154)", () => {
-    const gate: NodeDef = {
-      id: "gate", name: "gate", type: "doc-only",
-      inputs: [{ name: "in", repeated: false, side: "left" }],
-      outputs: [{ name: "out", repeated: false, side: "right" }],
-      interactive: false, view: { x: 200, y: 0 },
-    };
-    const yaml = serializePipeline(
-      makeFullPipeline([gate], [
-        {
-          source: { node: "gate", port: "out" },
-          target: { node: "end", port: "result" },
-          mode: "manual",
-          waypoints: [
-            { x: 120, y: 40 },
-            { x: 120, y: 220 },
-          ],
-        },
-      ]),
-    );
-    expect(yaml).toContain("mode: manual");
-    expect(yaml).toContain("waypoints:");
-    // The coordinates survive so the route travels with a shared pipeline.
-    expect(yaml).toContain("x: 120");
-    expect(yaml).toContain("y: 40");
-    expect(yaml).toContain("y: 220");
-  });
-
-  it("omits routing fields for an auto edge (no waypoints stored, #154)", () => {
-    const gate: NodeDef = {
-      id: "gate", name: "gate", type: "doc-only",
-      inputs: [{ name: "in", repeated: false, side: "left" }],
-      outputs: [{ name: "out", repeated: false, side: "right" }],
-      interactive: false, view: { x: 200, y: 0 },
-    };
-    const yaml = serializePipeline(
-      makeFullPipeline([gate], [
-        {
-          source: { node: "gate", port: "out" },
-          target: { node: "end", port: "result" },
-          mode: "auto",
-        },
-      ]),
-    );
-    // Auto edges recompute deterministically — nothing routing-related persists.
-    expect(yaml).not.toContain("mode:");
-    expect(yaml).not.toContain("waypoints:");
-  });
-
-  it("serializes an edge's target_side so the drop-position anchor survives reload (#168)", () => {
-    const impl: NodeDef = {
-      id: "impl", name: "impl", type: "code-mutating",
-      inputs: [], outputs: [{ name: "out", repeated: false, side: "right" }],
-      interactive: false, view: { x: 200, y: 0 },
-    };
-    const yaml = serializePipeline(
-      makeFullPipeline([impl], [
-        {
-          source: { node: "start", port: "user_prompt" },
-          target: { node: "impl", port: "user_prompt" },
-          target_side: "top",
-        },
-      ]),
-    );
-    expect(yaml).toContain("target_side: top");
-  });
-
-  it("omits target_side for a left-anchored (legacy) edge (#168)", () => {
-    const impl: NodeDef = {
-      id: "impl", name: "impl", type: "code-mutating",
-      inputs: [], outputs: [{ name: "out", repeated: false, side: "right" }],
-      interactive: false, view: { x: 200, y: 0 },
-    };
-    const yaml = serializePipeline(
-      makeFullPipeline([impl], [
-        {
-          source: { node: "start", port: "user_prompt" },
-          target: { node: "impl", port: "user_prompt" },
-        },
-      ]),
-    );
-    expect(yaml).not.toContain("target_side:");
-  });
-
-  it("serializes multi-field frontmatter with all fields at same depth", () => {
-    const node: NodeDef = {
-      id: "multi", name: "multi", type: "doc-only",
-      inputs: [{ name: "in", repeated: false }],
-      outputs: [{
-        name: "out", repeated: false,
-        frontmatter: {
-          verdict: { type: "enum", allowed: ["PASS", "FAIL"] },
-          score: { type: "int" },
-          summary: { type: "string" },
-        },
-      }],
-      interactive: false, view: { x: 200, y: 0 },
-    };
-    const yaml = serializePipeline(makeFullPipeline([node]));
-
-    const lines = yaml.split("\n");
-    const verdictLine = lines.find((l) => /^\s+verdict:/.test(l));
-    const scoreLine = lines.find((l) => /^\s+score:/.test(l));
-    const summaryLine = lines.find((l) => /^\s+summary:/.test(l));
-
-    expect(verdictLine).toBeDefined();
-    expect(scoreLine).toBeDefined();
-    expect(summaryLine).toBeDefined();
-
-    const indent = (l: string) => l.match(/^(\s*)/)?.[1].length ?? -1;
-    expect(indent(verdictLine!)).toBe(indent(scoreLine!));
-    expect(indent(scoreLine!)).toBe(indent(summaryLine!));
-  });
-
-  it("serializes a deeply nested edge when clause with in-predicate correctly", () => {
-    const gate: NodeDef = {
-      id: "gate", name: "gate", type: "doc-only",
-      inputs: [{ name: "in", repeated: false, side: "left" }],
-      outputs: [{ name: "out", repeated: false, side: "right" }],
-      interactive: false, view: { x: 200, y: 0 },
-    };
-    const yaml = serializePipeline(
-      makeFullPipeline([gate], [
-        {
-          source: { node: "gate", port: "out" },
-          target: { node: "end", port: "result" },
-          when: { verdict: { in: ["PASS", "APPROVED"] } },
-        },
-      ]),
-    );
-
-    // The YAML must not contain excessive indentation (more than 16 spaces
-    // for any line would indicate double-indent bug)
-    const lines = yaml.split("\n");
-    for (const line of lines) {
-      const leadingSpaces = line.match(/^(\s*)/)?.[1].length ?? 0;
-      expect(leadingSpaces).toBeLessThan(16);
-    }
-  });
-
-  it("serializes pipeline with variables correctly", () => {
-    const pipeline: PipelineDef = {
-      name: "vars-test", version: "1.0",
-      variables: {
-        max_iter: { type: "int", default: 5 },
-        threshold: { type: "float", default: 0.8 },
-      },
-      nodes: [
-        {
-          id: "start", name: "Start", type: "start",
-          inputs: [], outputs: [{ name: "user_prompt", repeated: false, side: "right" }],
-          interactive: false, view: { x: 0, y: 0 },
-        },
-        {
-          id: "end", name: "End", type: "end",
-          inputs: [{ name: "result", repeated: false, side: "left" }], outputs: [],
-          interactive: false, view: { x: 400, y: 0 },
-        },
-      ],
-      edges: [],
-    };
-    const yaml = serializePipeline(pipeline);
-    expect(yaml).toContain("variables:");
-    expect(yaml).toContain("max_iter: 5");
-    expect(yaml).toContain("threshold: 0.8");
-  });
-});
-
-describe("serializePipeline persists edge when/else (ADR-0011)", () => {
-  function makeEdgePipeline(edges: EdgeDef[]): PipelineDef {
-    return {
-      name: "edge-when-test",
-      version: "1.0",
-      variables: {},
-      nodes: [
-        {
-          id: "reviewer", name: "reviewer", type: "doc-only",
-          inputs: [{ name: "task", repeated: false, side: "left" }],
-          outputs: [{ name: "verdict", repeated: false, side: "right" }],
-          interactive: false, view: { x: 0, y: 0 },
-        },
-        {
-          id: "impl", name: "impl", type: "code-mutating",
-          inputs: [{ name: "review", repeated: false, side: "left" }],
-          outputs: [{ name: "diff", repeated: false, side: "right" }],
-          interactive: false, view: { x: 200, y: 0 },
-        },
-      ],
-      edges,
-    };
-  }
-
-  it("emits the when clause on a guarded edge", () => {
-    const yaml = serializePipeline(
-      makeEdgePipeline([
-        {
-          source: { node: "reviewer", port: "verdict" },
-          target: { node: "impl", port: "review" },
-          when: { verdict: { eq: "FAIL" } },
-        },
-      ]),
-    );
-    expect(yaml).toContain("when:");
-    expect(yaml).toContain("verdict:");
-    expect(yaml).toContain("eq: FAIL");
-  });
-
-  it("emits a canonical boolean (not a string) for a bool when value", () => {
-    const yaml = serializePipeline(
-      makeEdgePipeline([
-        {
-          source: { node: "reviewer", port: "verdict" },
-          target: { node: "impl", port: "review" },
-          when: { is_blocking: { eq: true } },
-        },
-      ]),
-    );
-    // The value must be a YAML boolean `true`, never the string "true".
-    expect(yaml).toMatch(/eq: true\b/);
-    expect(yaml).not.toContain('eq: "true"');
-  });
-
-  it("emits else: true on a fallback edge", () => {
-    const yaml = serializePipeline(
-      makeEdgePipeline([
-        {
-          source: { node: "reviewer", port: "verdict" },
-          target: { node: "impl", port: "review" },
-          else: true,
-        },
-      ]),
-    );
-    expect(yaml).toContain("else: true");
-  });
-
-  it("omits when/else on an unconditional edge", () => {
-    const yaml = serializePipeline(
-      makeEdgePipeline([
-        {
-          source: { node: "reviewer", port: "verdict" },
-          target: { node: "impl", port: "review" },
-        },
-      ]),
-    );
-    expect(yaml).not.toContain("when:");
-    expect(yaml).not.toContain("else:");
-  });
-});
-
-describe("serializePipeline persists port_type", () => {
-  function makePipelineWithTypedPorts(): PipelineDef {
-    const tester: NodeDef = {
-      id: "9NOnrpKY",
-      name: "Tester",
-      type: "doc-only",
-      inputs: [
-        { name: "screens", repeated: false, side: "left", port_type: "image_list" },
-      ],
-      outputs: [
-        { name: "screens-fixed", repeated: false, side: "right", port_type: "image_list" },
-        { name: "report", repeated: false, side: "right" },
-      ],
-      interactive: false,
-      view: { x: 200, y: 0 },
-    };
-    return {
-      name: "typed-ports-test",
-      version: "1.0",
-      variables: {},
-      nodes: [tester],
-      edges: [],
-    };
-  }
-
-  it("emits port_type: image_list for both input and output ports", () => {
-    const yaml = serializePipeline(makePipelineWithTypedPorts());
-    const occurrences = yaml.match(/port_type: image_list/g) ?? [];
-    // One for the input port (screens), one for the output port (screens-fixed).
-    expect(occurrences.length).toBe(2);
-  });
-
-  it("does not emit port_type for the default markdown type", () => {
-    const yaml = serializePipeline(makePipelineWithTypedPorts());
-    // The "report" output has no port_type set, so it must default to markdown
-    // implicitly and never appear in the YAML.
-    expect(yaml).not.toContain("port_type: markdown");
-  });
-
-  // #333: an `html` output port round-trips as `port_type: html` (emitted only
-  // because it is non-default), so a saved pipeline preserves it.
-  it("emits port_type: html for an html output port", () => {
-    const designer: NodeDef = {
-      id: "designer0",
-      name: "Designer",
-      type: "doc-only",
-      inputs: [],
-      outputs: [
-        { name: "report", repeated: false, side: "right", port_type: "html" },
-        { name: "notes", repeated: false, side: "right" },
-      ],
-      interactive: false,
-      view: { x: 0, y: 0 },
-    };
-    const yaml = serializePipeline({
-      name: "html-port-test",
-      version: "1.0",
-      variables: {},
-      nodes: [designer],
-      edges: [],
-    });
-    const occurrences = yaml.match(/port_type: html/g) ?? [];
-    expect(occurrences.length).toBe(1);
-    // The default-markdown "notes" port never carries a port_type.
-    expect(yaml).not.toContain("port_type: markdown");
-  });
-});
-
 describe("updateNode propagates port changes to edges", () => {
   it("renames edge source port when an output port is renamed", () => {
     const nodeA = makeNode({
@@ -2705,82 +2225,5 @@ describe("single-tab mode (#342)", () => {
       expect(useEditStore.getState().openTabs.map((t) => t.id)).toEqual(["a"]);
       expect(useEditStore.getState().pendingSingleTab).toBeNull();
     });
-  });
-});
-
-describe("exportNodeAsYaml (#345)", () => {
-  function node(overrides: Partial<NodeDef> = {}): NodeDef {
-    return {
-      id: "abc12345",
-      name: "Reviewer",
-      type: "doc-only",
-      inputs: [],
-      outputs: [{ name: "review", repeated: false, side: "right" }],
-      interactive: false,
-      view: { x: 42, y: 99 },
-      ...overrides,
-    };
-  }
-
-  it("emits a multi-line prompt as a block scalar, not a JSON-escaped single line", () => {
-    const yaml = exportNodeAsYaml(node(), "Line one.\nLine two.");
-    expect(yaml).toContain("prompt: |");
-    expect(yaml).toContain("  Line one.");
-    expect(yaml).toContain("  Line two.");
-    // The naive `dumpYaml` path would emit "Line one.\nLine two." — assert we don't.
-    expect(yaml).not.toContain('"Line one.\\nLine two."');
-  });
-
-  it("preserves a trailing newline via clip + a physical trailing newline (#345 round-trip)", () => {
-    const yaml = exportNodeAsYaml(node(), "Line one.\nLine two.\n");
-    // Clip (`|2`, not `|2-`) is chosen …
-    expect(yaml).toContain("prompt: |2\n");
-    expect(yaml).not.toContain("prompt: |2-");
-    // … and the source must actually end in a newline for clip to keep it.
-    // Without this the trailing `\n` is silently dropped on round-trip.
-    expect(yaml.endsWith("  Line two.\n")).toBe(true);
-  });
-
-  it("strips (no phantom newline) when the prompt has no trailing newline", () => {
-    const yaml = exportNodeAsYaml(node(), "Line one.\nLine two.");
-    expect(yaml).toContain("prompt: |2-\n");
-    expect(yaml.endsWith("  Line two.")).toBe(true);
-    expect(yaml.endsWith("\n")).toBe(false);
-  });
-
-  it("omits id, view, and any edges (a node carries none)", () => {
-    const yaml = exportNodeAsYaml(node(), "p");
-    expect(yaml).not.toMatch(/(^|\n)id:/);
-    expect(yaml).not.toContain("view:");
-    expect(yaml).not.toContain("abc12345");
-    expect(yaml).not.toContain("edges");
-  });
-
-  it("includes model when set and omits it when unset (#296)", () => {
-    expect(exportNodeAsYaml(node({ model: "opus" }), "p")).toContain("model: opus");
-    expect(exportNodeAsYaml(node({ model: null }), "p")).not.toContain("model:");
-  });
-
-  it("emits interactive only when true", () => {
-    expect(exportNodeAsYaml(node({ interactive: true }), "p")).toContain("interactive: true");
-    expect(exportNodeAsYaml(node({ interactive: false }), "p")).not.toContain("interactive:");
-  });
-
-  it("omits a port's default side but keeps a non-default one", () => {
-    // right is the default output side → omitted for a clean, minimal YAML.
-    expect(exportNodeAsYaml(node(), "p")).not.toContain("side:");
-    const topped = exportNodeAsYaml(
-      node({ outputs: [{ name: "review", repeated: false, side: "top" }] }),
-      "p",
-    );
-    expect(topped).toContain("side: top");
-  });
-
-  it("emits an empty prompt as an explicit empty string", () => {
-    expect(exportNodeAsYaml(node(), "")).toContain('prompt: ""');
-  });
-
-  it("is library-entry-shaped: name + type at the root", () => {
-    expect(exportNodeAsYaml(node(), "p").startsWith("name: Reviewer\ntype: doc-only")).toBe(true);
   });
 });
